requirements.txt
PyQt5>=5.15.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.9.0
tqdm>=4.62.0
